pyproject.toml
[tool.poetry]
name = "pypi_scout"
version = "0.0.1"
description = "PyPI Scout helps you find PyPI packages through natural language prompts with the help of Large Language Models (LLMs)."
authors = ["Florian Maas <ffpgmaas@gmail.com>"]
repository = "https://github.com/fpgmaas/pypi-scout"
documentation = "https://fpgmaas.github.io/pypi-scout/"
readme = "README.md"
packages = [
  {include = "pypi_scout"}
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
beautifulsoup4 = "^4.12.3"
polars = "^0.20.31"
sentence-transformers = "^3.0.1"
lxml = "^5.2.2"
python-dotenv = "^1.0.1"
pinecone = "^4.0.0"
tqdm = "^4.66.4"
fastapi = "^0.111.0"
pydantic = "^2.7.4"
uvicorn = "^0.30.1"
gdown = "^5.2.0"
azure-storage-blob = "^12.20.0"
slowapi = "^0.1.9"
starlette = "^0.37.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.14.0"
deptry = "^0.12.0"
pre-commit = "^3.4.0"
tox = "^4.11.1"
jupyterlab = "^4.2.2"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = {extras = ["python"], version = "^0.23.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["pypi_scout"]


[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.deptry]
extend_exclude = [
    "frontend"
]

[tool.deptry.per_rule_ignores]
DEP002 = ["lxml", "uvicorn"]
